+++ ppcls/data/preprocess/ops/operators.py
@@ -269,7 +269,7 @@ class CropImage(object):
 
         w_end = w_start + w
         h_end = h_start + h
-        return img[h_start:h_end, w_start:w_end, :]
+        return img[h_start:h_end, w_start:w_end, :].astype('float32')
 
 
 class Padv2(object):
@@ -517,7 +517,6 @@ class NormalizeImage(object):
         from PIL import Image
         if isinstance(img, Image.Image):
             img = np.array(img)
-
         assert isinstance(img,
                           np.ndarray), "invalid input 'img' in NormalizeImage"
